requirements.txt
flake8==3.9.0
pytest==6.2.5
pre-commit==2.15.0
wemake-python-styleguide==0.15.3
flake8-bandit==2.1.2
bandit==1.7.2
isort==5.10.1
Faker==15.1.0
psycopg2-binary==2.9.4
Django==4.1.2
python_dotenv==0.21.0
django_split_settings==1.2.0
django-debug-toolbar==3.7.0
pydantic==1.10.2
tabulate==0.9.0
